chore: move GiftHistory to shared
